praetor/requirements.txt
# PRAETOR Enterprise — Python dependencies
# Requires Python 3.10+ (uses X | Y union syntax)
# Install: pip install -r requirements.txt

# Web framework & server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# HTTP client
httpx>=0.27.0

# Data validation
pydantic>=2.6.0

# Redis (async)
redis>=5.0.0

# YAML config loading
pyyaml>=6.0.1

# Anthropic SDK (agent LLM calls)
anthropic>=0.25.0

# OpenTelemetry
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0

# Prometheus metrics
prometheus-client>=0.20.0
